Move config setup out of nu_cli
