+++ crates/nu-cli/src/commands.rs
@@ -1,11 +1,12 @@
-use crate::util::{gather_parent_env_vars, report_error};
+use crate::util::report_error;
 use log::info;
 use miette::Result;
 use nu_engine::{convert_env_values, eval_block};
 use nu_parser::{parse, trim_quotes};
+use nu_protocol::engine::Stack;
 use nu_protocol::{
     engine::{EngineState, StateDelta, StateWorkingSet},
-    Config, PipelineData, Span, Spanned, Value, CONFIG_VARIABLE_ID,
+    Config, PipelineData, Spanned,
 };
 use std::path::Path;
 
@@ -13,15 +14,10 @@ pub fn evaluate_commands(
     commands: &Spanned<String>,
     init_cwd: &Path,
     engine_state: &mut EngineState,
+    stack: &mut Stack,
     input: PipelineData,
     is_perf_true: bool,
 ) -> Result<()> {
-    #[cfg(feature = "plugin")]
-    crate::config_files::add_plugin_file(engine_state);
-
-    // First, set up env vars as strings only
-    gather_parent_env_vars(engine_state);
-
     // Run a command (or commands) given to us by the user
     let (block, delta) = {
         let mut working_set = StateWorkingSet::new(engine_state);
@@ -50,18 +46,6 @@ pub fn evaluate_commands(
         report_error(&working_set, &err);
     }
 
-    let mut stack = nu_protocol::engine::Stack::new();
-
-    // Set up our initial config to start from
-    stack.vars.insert(
-        CONFIG_VARIABLE_ID,
-        Value::Record {
-            cols: vec![],
-            vals: vec![],
-            span: Span { start: 0, end: 0 },
-        },
-    );
-
     let config = match stack.get_config() {
         Ok(config) => config,
         Err(e) => {
@@ -73,13 +57,13 @@ pub fn evaluate_commands(
     };
 
     // Make a note of the exceptions we see for externals that look like math expressions
-    let exceptions = crate::util::external_exceptions(engine_state, &stack);
+    let exceptions = crate::util::external_exceptions(engine_state, stack);
     engine_state.external_exceptions = exceptions;
 
     // Merge the delta in case env vars changed in the config
-    match nu_engine::env::current_dir(engine_state, &stack) {
+    match nu_engine::env::current_dir(engine_state, stack) {
         Ok(cwd) => {
-            if let Err(e) = engine_state.merge_delta(StateDelta::new(), Some(&mut stack), cwd) {
+            if let Err(e) = engine_state.merge_delta(StateDelta::new(), Some(stack), cwd) {
                 let working_set = StateWorkingSet::new(engine_state);
                 report_error(&working_set, &e);
                 std::process::exit(1);
@@ -93,15 +77,15 @@ pub fn evaluate_commands(
     }
 
     // Translate environment variables from Strings to Values
-    if let Some(e) = convert_env_values(engine_state, &stack) {
+    if let Some(e) = convert_env_values(engine_state, stack) {
         let working_set = StateWorkingSet::new(engine_state);
         report_error(&working_set, &e);
         std::process::exit(1);
     }
 
-    match eval_block(engine_state, &mut stack, &block, input, false, false) {
+    match eval_block(engine_state, stack, &block, input, false, false) {
         Ok(pipeline_data) => {
-            crate::eval_file::print_table_or_error(engine_state, &mut stack, pipeline_data, &config)
+            crate::eval_file::print_table_or_error(engine_state, stack, pipeline_data, &config)
         }
         Err(err) => {
             let working_set = StateWorkingSet::new(engine_state);
