+++ crates/nu-cli/src/lib.rs
@@ -15,6 +15,7 @@ mod validation;
 
 pub use commands::evaluate_commands;
 pub use completions::NuCompleter;
+pub use config_files::eval_config_contents;
 pub use errors::CliError;
 pub use eval_file::evaluate_file;
 pub use nu_highlight::NuHighlight;
@@ -23,5 +24,10 @@ pub use prompt::NushellPrompt;
 pub use repl::evaluate_repl;
 pub use syntax_highlight::NuHighlighter;
 pub use util::print_pipeline_data;
-pub use util::{get_init_cwd, report_error};
+pub use util::{eval_source, gather_parent_env_vars, get_init_cwd, report_error};
 pub use validation::NuValidator;
+
+#[cfg(feature = "plugin")]
+pub use config_files::add_plugin_file;
+#[cfg(feature = "plugin")]
+pub use config_files::read_plugin_file;
