+++ crates/nu-cli/src/repl.rs
@@ -1,8 +1,8 @@
 use crate::reedline_config::{add_completion_menu, add_history_menu};
-use crate::{config_files, prompt_update, reedline_config};
+use crate::{prompt_update, reedline_config};
 use crate::{
     reedline_config::KeybindingsMode,
-    util::{eval_source, gather_parent_env_vars, report_error},
+    util::{eval_source, report_error},
 };
 use crate::{NuCompleter, NuHighlighter, NuValidator, NushellPrompt};
 use log::info;
@@ -11,17 +11,20 @@ use miette::{IntoDiagnostic, Result};
 use nu_color_config::get_color_config;
 use nu_engine::convert_env_values;
 use nu_parser::lex;
+use nu_protocol::engine::Stack;
+use nu_protocol::PipelineData;
 use nu_protocol::{
     engine::{EngineState, StateWorkingSet},
     Config, ShellError, Span, Value, CONFIG_VARIABLE_ID,
 };
-use nu_protocol::{PipelineData, Spanned};
 use reedline::{DefaultHinter, Emacs, Vi};
+use std::path::PathBuf;
 use std::{sync::atomic::Ordering, time::Instant};
 
 pub fn evaluate_repl(
     engine_state: &mut EngineState,
-    config_file: Option<Spanned<String>>,
+    stack: &mut Stack,
+    history_path: Option<PathBuf>,
     is_perf_true: bool,
 ) -> Result<()> {
     // use crate::logger::{configure, logger};
@@ -30,34 +33,34 @@ pub fn evaluate_repl(
     let mut entry_num = 0;
 
     let mut nu_prompt = NushellPrompt::new();
-    let mut stack = nu_protocol::engine::Stack::new();
+    // let mut stack = nu_protocol::engine::Stack::new();
 
     // First, set up env vars as strings only
-    gather_parent_env_vars(engine_state);
+    // gather_parent_env_vars(engine_state);
 
     // Set up our initial config to start from
-    stack.vars.insert(
-        CONFIG_VARIABLE_ID,
-        Value::Record {
-            cols: vec![],
-            vals: vec![],
-            span: Span::new(0, 0),
-        },
-    );
+    // stack.vars.insert(
+    //     CONFIG_VARIABLE_ID,
+    //     Value::Record {
+    //         cols: vec![],
+    //         vals: vec![],
+    //         span: Span::new(0, 0),
+    //     },
+    // );
 
     if is_perf_true {
         info!("read_plugin_file {}:{}:{}", file!(), line!(), column!());
     }
 
-    #[cfg(feature = "plugin")]
-    config_files::read_plugin_file(engine_state, &mut stack, is_perf_true);
-
-    if is_perf_true {
-        info!("read_config_file {}:{}:{}", file!(), line!(), column!());
-    }
-
-    config_files::read_config_file(engine_state, &mut stack, config_file, is_perf_true);
-    let history_path = config_files::create_history_path();
+    // #[cfg(feature = "plugin")]
+    // config_files::read_plugin_file(engine_state, &mut stack, is_perf_true);
+    //
+    // if is_perf_true {
+    //     info!("read_config_file {}:{}:{}", file!(), line!(), column!());
+    // }
+    //
+    // config_files::read_config_file(engine_state, &mut stack, config_file, is_perf_true);
+    // let history_path = config_files::create_history_path();
 
     // logger(|builder| {
     //     configure(&config.log_level, builder)?;
@@ -77,13 +80,13 @@ pub fn evaluate_repl(
     }
 
     // Translate environment variables from Strings to Values
-    if let Some(e) = convert_env_values(engine_state, &stack) {
+    if let Some(e) = convert_env_values(engine_state, stack) {
         let working_set = StateWorkingSet::new(engine_state);
         report_error(&working_set, &e);
     }
 
     // Make a note of the exceptions we see for externals that look like math expressions
-    let exceptions = crate::util::external_exceptions(engine_state, &stack);
+    let exceptions = crate::util::external_exceptions(engine_state, stack);
     engine_state.external_exceptions = exceptions;
 
     // seed env vars
@@ -229,7 +232,7 @@ pub fn evaluate_repl(
         let prompt = prompt_update::update_prompt(
             &config,
             engine_state,
-            &stack,
+            stack,
             &mut nu_prompt,
             is_perf_true,
         );
@@ -251,7 +254,7 @@ pub fn evaluate_repl(
                 let start_time = Instant::now();
                 let tokens = lex(s.as_bytes(), 0, &[], &[], false);
                 // Check if this is a single call to a directory, if so auto-cd
-                let cwd = nu_engine::env::current_dir_str(engine_state, &stack)?;
+                let cwd = nu_engine::env::current_dir_str(engine_state, stack)?;
                 let path = nu_path::expand_path_with(&s, &cwd);
 
                 let orig = s.clone();
@@ -314,7 +317,7 @@ pub fn evaluate_repl(
 
                     eval_source(
                         engine_state,
-                        &mut stack,
+                        stack,
                         s.as_bytes(),
                         &format!("entry #{}", entry_num),
                         PipelineData::new(Span::new(0, 0)),
@@ -337,7 +340,7 @@ pub fn evaluate_repl(
                 }
 
                 // Make a note of the exceptions we see for externals that look like math expressions
-                let exceptions = crate::util::external_exceptions(engine_state, &stack);
+                let exceptions = crate::util::external_exceptions(engine_state, stack);
                 engine_state.external_exceptions = exceptions;
             }
             Ok(Signal::CtrlC) => {
